day 9

+++ src/main/scala/com/cberes/advent_of_code/Day07.scala
@@ -12,17 +12,17 @@ import scala.concurrent.duration._
 
 object Day07 {
   @tailrec
-  def solve(state : State, input : List[Int], previousOutput : Int = 0) : Int = input match {
+  def solve(state : State, input : List[Int], previousOutput : Long = 0) : Long = input match {
     case phaseSetting :: rest =>
       val output = computeSeries(state, List(phaseSetting, previousOutput))
       solve(state, rest, output.head)
     case Nil => previousOutput
   }
 
-  def maxOutput(state : State, n : Int) : Int = allInputCombinations(n).map(it => solve(state, it)).max
+  def maxOutput(state : State, n : Int) : Long = allInputCombinations(n).map(it => solve(state, it)).max
 
-  def solveFeedback(state : State, input : List[Int]) : Int = {
-    val inputQueues = input.map(_ => new LinkedBlockingQueue[Int]())
+  def solveFeedback(state : State, input : List[Int]) : Long = {
+    val inputQueues = input.map(_ => new LinkedBlockingQueue[Long]())
     input.indices.foreach(i => inputQueues(i).offer(input(i)))
     inputQueues.head.offer(0)
 
@@ -40,13 +40,13 @@ object Day07 {
     inputQueues.head.peek()
   }
 
-  def maxOutputFeedback(state : State, n : Int, minSetting : Int) : Int =
+  def maxOutputFeedback(state : State, n : Int, minSetting : Int) : Long =
     allInputCombinations(n, minSetting).map(it => solveFeedback(state, it)).max
 
-  def computeSeries(state : State, inputs : Seq[Int]) : Seq[Int] = {
-    val inputQueue = new mutable.Queue[Int]
+  def computeSeries(state : State, inputs : Seq[Long]) : Seq[Long] = {
+    val inputQueue = new mutable.Queue[Long]
     inputQueue.enqueueAll(inputs)
-    val outputs = new mutable.ListBuffer[Int]
+    val outputs = new mutable.ListBuffer[Long]
     compute(state, computer(inputQueue.dequeue, outputs.append))
     outputs.toSeq
   }
@@ -59,12 +59,12 @@ object Day07 {
       .filter(_.toSet.size == n)
   }
 
-  def part1(args : Array[String]) : Int = doWithLines(args.head) { lines =>
+  def part1(args : Array[String]) : Long = doWithLines(args.head) { lines =>
     val state = lines.next().toState
     maxOutput(state, 5)
   }
 
-  def part2(args : Array[String]) : Int = doWithLines(args.head) { lines =>
+  def part2(args : Array[String]) : Long = doWithLines(args.head) { lines =>
     val state = lines.next().toState
     maxOutputFeedback(state, 5, 5)
   }
